refactor(T11404): reclaim core/src/tools — relocate 5 file squatters to honest siblings (E3 · SG-PACKAGE-ARCH)
The atomic-tool namespace (core/src/tools) was squatted by non-tool modules.
Relocates the 5 FILE squatters to their honest, already-existing core homes:

- engine-ops.ts (1338 LOC) → core/src/engine/
- adr-backfill-walker.ts (765 LOC) → core/src/backfill/
- doctor-project.ts → core/src/doctor/
- scaffold-global.ts, scaffold-project.ts → core/src/scaffold/

Same-depth moves preserve each file's relative imports (../paths.js etc.);
tools/index.ts barrel repointed to the new locations + the one direct importer
(internal.ts) fixed → @cleocode/core public API is UNCHANGED (zero behavior
change). core/src/tools now holds only the atomic primitives (fs/shell/guard)
plus the 3 remaining DIR squatters (brain-tools/, task-tools/, sdk/), which
need depth-fixups and are a scoped follow-up — T11404 stays open.

Verified: core build (tsc) · cleo build (public API intact) · cleo check arch
5/5 · tools-vs-skills boundary + json-stream-hygiene clean · biome.
