+++ packages/core/src/doctor/doctor-project.ts
