+++ packages/core/src/scaffold/scaffold-global.ts
