+++ packages/core/src/tools/__tests__/project-tools.test.ts
@@ -13,8 +13,8 @@ import { mkdtempSync, rmSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { doctorProject } from '../doctor-project.js';
-import { scaffoldProject } from '../scaffold-project.js';
+import { doctorProject } from '../../doctor/doctor-project.js';
+import { scaffoldProject } from '../../scaffold/scaffold-project.js';
 
 // ── scaffold-project ──────────────────────────────────────────────────
